==> notebooks/01_extract_metadata.ipynb <==
## Get metadata from path and filename and export to file:
* path
* organism: E. faecium (Efcm), E.coli(Ecoli), S. aureus (Sau)
* sequencing technology: HS, MS, NX
* Read number: 1/2
* evaluation (label): good/ugly (1/0)

import os
import re
import pandas as pd

rootdir = 'training_data'

reads =	{
  "filename": [],
  "organism": [],
  "technology": [],
  "read_number": [],
  "evaluation": []
}

for root, dirs, files in os.walk(rootdir):
    if root.endswith("fastqc"):
        extracted_vars = root.split('/')
        reads["organism"].append(extracted_vars[-5])
        reads["evaluation"].append(extracted_vars[-4])
        reads["technology"].append(extracted_vars[-3])
        reads["filename"].append(extracted_vars[-1].replace("_fastqc", ""))
        read_number = extracted_vars[-1].split('_')[-3][1]
        reads["read_number"].append(read_number)

metadata = pd.DataFrame(reads)
metadata

metadata.to_json('exported_datasets/metadata.json')
