==> notebooks/02_extract_raw_fastq.ipynb <==
import os
import numpy as np
import pyfastx
import matplotlib.pyplot as plt
import pandas as pd

#https://biopython.org/docs/1.75/api/Bio.SeqIO.QualityIO.html
#http://biopython.org/DIST/docs/tutorial/Tutorial.html#sec79
# https://biopython.org/wiki/Split_large_file
# https://pypi.org/project/pyfastx/
# https://jakevdp.github.io/PythonDataScienceHandbook/01.07-timing-and-profiling.html


#examples:
# 69kB training_data/Sau/ugly/MS/180727-18-5425-18-01680_S8_L001_R1_001.fastq.gz
# 52MB training_data/Sau/ugly/MS/200204_20-00746_19-03927_S9_L000_R2_001.fastq.gz
# 250MB training_data/Sau/ugly/MS/200327_20-04029_20-00557_S26_L000_R2_001.fastq.gz
# 1.5GB training_data/Sau/ugly/MS/181002-18-6992-776-18_S2_L001_R2_001.fastq.gz

#### convert ascii quality with offset of 33 to phred score

def phred_score(quality):
    return ord(quality)-33

#### extract quality means and base contents per position

def extract_fastq_positions(path, max_len_estim=1000):
    phred_sum = max_len_estim*[0] # sum all qualities per position, initialize with enough space for max_len
    counts = max_len_estim*[0] # count occurences of each position
    base_content = {'G': max_len_estim*[0], #count occurences of each base per position
                    'C': max_len_estim*[0],
                    'A': max_len_estim*[0], 
                    'T': max_len_estim*[0],
                    'N': max_len_estim*[0]}
    # fill lists with values from fastq file
    for name, seq, qual in pyfastx.Fastq(path, build_index=False):
        for i in range(len(seq)):
            phred_sum[i] += phred_score(qual[i])
            counts[i] += 1
            base_content[seq[i]][i] += 1
    
    phred_sum = np.array(phred_sum)
    counts = np.array(counts)
    max_len = np.min(np.where(counts==0)) #find index of first 0 value as maximum length of sequences
    phred_sum = phred_sum[0:max_len]
    counts = counts[0:max_len]
    means = phred_sum/counts
    
    # convert the lists in base_content to np.arrays to perform calculations
    for key in base_content:
        base_content[key] = np.array(base_content[key])
    max_len = np.min(np.where(base_content['G']==0)) #find index of first 0 value as maximum length of sequences
    for key in base_content:
        base_content[key] = base_content[key][0:max_len] #remove unused positions
    #sum all occurences to calculate percentages
    sums = (base_content['G'] + base_content['C'] + base_content['T'] +
            base_content['A'] + base_content['N'])
    ncontent_total = base_content['N'].sum()/sums.sum()
    for key in base_content:
        base_content[key] = base_content[key]/sums
    
    return means, base_content, ncontent_total

#### import fastq files from directories and extract positional data

#from datetime import datetime #remove

def import_reads(rootdir):
    filenames = []
    all_phred_means = []
    all_base_content = []
    all_n_content = []
    for root, dirs, files in os.walk(rootdir):
        for name in files:
            filepath = root + os.sep + name
            if filepath.endswith(".fastq.gz"):
                phred_means, base_content, n_content = extract_fastq_positions(filepath)
                all_phred_means.append(phred_means)
                all_base_content.append(base_content)
                all_n_content.append(n_content)
                filenames.append(name.replace(".fastq.gz", ""))
                #print("finished " + filepath + " " + str(datetime.now(tz=None)))
    
    fastq_positions = pd.DataFrame(all_base_content, columns=['G', 'C', 'A', 'T', 'N'])
    fastq_positions['filename']= filenames
    fastq_positions['phred_means'] = all_phred_means
    fastq_positions['n_content'] = all_n_content
    
    return fastq_positions

# runtime ~4hrs for 184 files
position_data = import_reads('training_data')

position_data

#### export dataset

position_data.to_json('exported_datasets/from_fastq_raw.json')
