==> notebooks/03_fastqc_data_extraction_and_merge.ipynb <==
# Data Extraction from FastQC Results

#### Import fastqQCParser   
https://pypi.org/project/fastqcparser/#description

from pprint import pprint
from fastqcparser import FastQCParser

import pandas as pd
import os

#### import all fastqc results from directories

rootdir = 'training_data'

def import_reads(rootdir):
    reads = []
    for root, dirs, files in os.walk(rootdir):
        for name in files:
            filepath = root + os.sep + name
            if filepath.endswith("fastqc_data.txt"):
                reads.append(FastQCParser(filepath))
    return reads

reads = import_reads(rootdir)

len(reads)

#### extract module names list (for column titles)

module_list = list(reads[0].modules.keys())
# imported reads don't include the module 'Overrepresented sequences' 
# if there are none in the read, so we manually add the module to the list
if(len(module_list) < 11):
    module_list.insert(9, 'Overrepresented sequences')
module_list.append('Module Statuses')
module_list.append('filename')
print(module_list)

#### fill columns for one read

def create_single_read_dataframe(result):
    module_result = []
    module_status = []
    
    for module in result.modules:
        result_data = pd.DataFrame(result.modules[module]['data'])
        result_data.columns = result.modules[module]['fieldnames']
        module_status.append(result.modules[module]['status'])
        module_result.append(result_data)
    # imported reads don't include the module 'Overrepresented sequences' 
    # if there are none in the read, so we manually add the status list
    # and an empty dataframe /is None better?
    if len(module_status) < 11:
        module_status.insert(9, 'pass')
        module_result.insert(9, pd.DataFrame())
    module_result.append(module_status)
    module_result.append(module_result[0].Value[0].replace(".fastq.gz", ""))
    module_series = pd.Series(data=module_result, index=module_list)
    single_read = module_series.to_frame().T#.set_index('filename')
    return single_read

#### apply to all reads and create single dataframe with all results

read_list = []
for read in reads:
    single_read = create_single_read_dataframe(read)
    read_list.append(single_read)

read_results = pd.concat(read_list)

type(read_results['Basic Statistics'])

#### import metadata

metadata = pd.read_json('exported_datasets/metadata.json')

#### merge both datasets

complete_reads = pd.merge(read_results, metadata, on='filename', how='inner').set_index('filename')
complete_reads.head()

#### import positional data from raw fastq dataset

position_data = pd.read_json('exported_datasets/from_fastq_raw.json')

# convert lists to np.arrays (not necessary if dataframe only used for export)
import numpy as np
for category in position_data:
    if type(position_data[category][0]) is list:
        position_data[category] = position_data[category].apply(np.array)
        print(category, 'converted to np.array')

position_data.set_index('filename', inplace=True)

#### add positional dataset to complete dataset

complete_reads = complete_reads.join(position_data)

complete_reads.head()

#### export complete dataset

complete_reads.to_json('exported_datasets/complete_set.json')
print('exported complete dataset')
