==> notebooks/05_data_preparation_ml.ipynb <==
## Prepare data for use in machine learning algorithms
* extract features from nested dataframes  
* convert features to numeric types

import pandas as pd
import numpy as np

#### import dataset

ngs_reads = pd.read_json('exported_datasets/complete_set.json')

# convert dict type columns to dataframes, lists to np arrays
# not necessary if those columns are only used for export
for category in ngs_reads:
    if type(ngs_reads[category][0]) is dict:
        ngs_reads[category] = ngs_reads[category].apply(pd.DataFrame)
        print(category, 'converted to DF')
    elif type(ngs_reads[category][0]) is list:
        ngs_reads[category] = ngs_reads[category].apply(np.array)
        print(category, 'converted to np.array')

ngs_reads.reset_index(drop=True, inplace=True)

### Basic Statistics
#### extract features from Basic Statistics

ngs_reads['Basic Statistics'][0]

# extract features from Basic Statistics

total_sequences = []
percent_gc = []
min_sequence_length = []
max_sequence_length = []
#(ngs_reads['Basic Statistics'][i]['Value'][4]) # sequences flagged as poor quality - 0 for all data points

for i in range(ngs_reads.shape[0]):
    total_sequences.append(ngs_reads['Basic Statistics'][i]['Value'][3])
    percent_gc.append(ngs_reads['Basic Statistics'][i]['Value'][6])
    length_min_max = str(ngs_reads['Basic Statistics'][i]['Value'][5]).split('-')
    min_sequence_length.append(length_min_max[0])
    max_sequence_length.append(length_min_max[-1])
total_sequences = np.asarray(total_sequences, dtype=np.int64)
percent_gc = np.asarray(percent_gc, dtype=np.int64)
min_sequence_length = np.asarray(min_sequence_length, dtype=np.int64)
max_sequence_length = np.asarray(max_sequence_length, dtype=np.int64)
ngs_reads['total_sequences'] = total_sequences
ngs_reads['percent_gc'] = percent_gc
ngs_reads['min_sequence_length'] = min_sequence_length
ngs_reads['max_sequence_length'] = max_sequence_length

ngs_reads.drop(columns='Basic Statistics', inplace=True)

### Evaluation (target variable)
#### convert evaluation value to numeric type

ngs_reads['evaluation'].replace({'ugly': 0, 'good': 1}, inplace=True)

### Modul Statuses
#### convert statuses to numeric type

status_replacements = {'fail':0, 'warn':1, 'pass':2}

for i in range(ngs_reads.shape[0]):
    statuses_list = [status_replacements.get(n, n) for n in ngs_reads['Module Statuses'].iat[i]]
    ngs_reads['Module Statuses'].iat[i] = statuses_list

#### add inner array data from Module Statuses as df columns

def create_status_column_name(module_nr):
    return 'module_' + str(module_nr) + '_status'

column_names_status = list(map(create_status_column_name, range(len(ngs_reads['Module Statuses'][0]))))

statuses_df = pd.DataFrame(ngs_reads['Module Statuses'].to_list(), columns=column_names_status)
ngs_reads = ngs_reads.join(statuses_df)
ngs_reads.drop(columns='Module Statuses', inplace=True)

#### (Todo: encode technology?)

#### remove status of module 0  
Module 0 is Basic Statistics and always has the same value (pass/2)

# number of unique values in column module_0_status
len(ngs_reads['module_0_status'].unique())

ngs_reads.drop('module_0_status', axis=1, inplace=True)

#### export complete prepared dataset

print('export full dataset')
print('shape -', ngs_reads.shape)
ngs_reads.to_json('exported_datasets/prepared_dataset.json')

#### remove module columns for export to to simple organism datasets

cols = ['Per base sequence quality', 'Per tile sequence quality',
       'Per sequence quality scores', 'Per base sequence content',
       'Per sequence GC content', 'Per base N content',
       'Sequence Length Distribution', 'Sequence Duplication Levels',
       'Overrepresented sequences', 'Adapter Content', 'G', 'C', 'A',
       'T', 'N', 'phred_means', 'n_content']
ngs_reads_simple = ngs_reads.drop(ngs_reads[cols], axis=1)
ngs_reads_simple.head()

#### split dataset by organism and export dataset for each

grouped = ngs_reads_simple.groupby(ngs_reads.organism)
efcm = grouped.get_group('Efcm')
efcm.reset_index(drop=True)
sau = grouped.get_group('Sau')
sau.reset_index(drop=True)
ecoli = grouped.get_group('Ecoli')
ecoli.reset_index(drop=True)

print('exported datasets')
print('all -', ngs_reads_simple.shape)
ngs_reads_simple.to_json('exported_datasets/all_simple.json')
print('efcm -', efcm.shape)
efcm.to_json('exported_datasets/efcm_simple.json')
print('sau -', sau.shape)
sau.to_json('exported_datasets/sau_simple.json')
print('ecoli -', ecoli.shape)
ecoli.to_json('exported_datasets/ecoli_simple.json')
